Everywhere: Use svset when collecting list of globals from diverse sources.
This saves quite a few duplicates in practice, and better matches
the desired operation (the comments already say "compute the set", etc).

Leave the old "addFullFunctionList" interface around to ensure not to break
things like SAFECode.  However, the implementation is now simply a wrapper
around the set version and as such should probably be converted at some point.